full refactor of project to client-server
In anticipation of testing, the project has been refactored to a client-server structure similar to that suggested here: https://github.com/angular-app/angular-app and that generated by https://github.com/angular-fullstack/generator-angular-fullstack where each module is a block of logic: markup-logic-test and possibly style. In addition to the refactoring a beautifier was run on each file to ensure consistent style going forward.